move get-block-from-file
